Support for live theme changes
Don't store theme settings between renders (still put them in settings, but clear and reload when we re-render).
Remove remaining references to global "colorStyles" object (always use settings)
Calculate `interpolator` from gradient as required, rather than all the time.
