+++ packages/perspective-viewer-d3fc/src/js/plugin/template.js
@@ -67,11 +67,10 @@ class D3FCChartElement extends HTMLElement {
         if (oldSettings) {
             const oldValues = [oldSettings.crossValues, oldSettings.mainValues, oldSettings.splitValues];
             const newValues = [newSettings.crossValues, newSettings.mainValues, newSettings.splitValues];
-            if (areArraysEqualSimple(oldValues, newValues)) return {...oldSettings, data: newSettings.data};
+            if (areArraysEqualSimple(oldValues, newValues)) return {...oldSettings, data: newSettings.data, colorStyles: null};
         }
         this.remove();
-        const {colorStyles} = oldSettings || {};
-        return {...newSettings, colorStyles};
+        return newSettings;
     }
 }
