+++ packages/perspective-viewer-d3fc/src/js/series/seriesRange.js
@@ -23,16 +23,38 @@ export function seriesColorRange(settings, data, valueName, customExtent) {
         domain()
             .valueName(valueName)
             .pad([0, 0])(data);
-    let interpolator = settings.colorStyles.interpolator.full;
+    let gradient = settings.colorStyles.gradient.full;
 
     if (extent[0] >= 0) {
-        interpolator = settings.colorStyles.interpolator.positive;
+        gradient = settings.colorStyles.gradient.positive;
     } else if (extent[1] <= 0) {
-        interpolator = settings.colorStyles.interpolator.negative;
+        gradient = settings.colorStyles.gradient.negative;
     } else {
         const maxVal = Math.max(-extent[0], extent[1]);
         extent = [-maxVal, maxVal];
     }
 
+    const interpolator = multiInterpolator(gradient);
     return d3.scaleSequential(interpolator).domain(extent);
 }
+
+const multiInterpolator = gradientPairs => {
+    // A new interpolator that calls through to a set of
+    // interpolators between each value/color pair
+    const interpolators = gradientPairs.slice(1).map((p, i) => d3.interpolate(gradientPairs[i][1], p[1]));
+    return value => {
+        const index = gradientPairs.findIndex((p, i) => i < gradientPairs.length - 1 && value <= gradientPairs[i + 1][0] && value > p[0]);
+        if (index === -1) {
+            if (value <= gradientPairs[0][0]) {
+                return gradientPairs[0][1];
+            }
+            return gradientPairs[gradientPairs.length - 1][1];
+        }
+
+        const interpolator = interpolators[index];
+        const [value1] = gradientPairs[index];
+        const [value2] = gradientPairs[index + 1];
+
+        return interpolator((value - value1) / (value2 - value1));
+    };
+};
